+++ src/market_data/ingest/mod.rs
@@ -1,4 +1,7 @@
 pub mod account_filter;
+pub mod account_handler;
+pub mod account_host;
+pub mod account_parse;
 pub mod host;
 pub mod tx_filter;
 pub mod tx_handler;
@@ -8,6 +11,13 @@ pub mod tx_parse;
 pub use account_filter::{
     account_geyser_dispatch_priority_high, account_geyser_update_is_dex_pool_owner,
     account_geyser_update_might_be_relevant,
+};
+pub use account_handler::handle_geyser_account_update;
+pub use account_host::{AccountBinArrayView, AccountIngestHost, AccountTrackedWalletView};
+pub use account_parse::{
+    try_parse_mint_account, try_parse_token_account_balance, wallet_geyser_snapshots_to_publish,
+    wsol_ata_balance_lamports_from_geyser_data, WalletGeyserSnapshotMint,
+    WalletGeyserSnapshotToPublish, WalletGeyserUpdateSource,
 };
 pub use host::IngestHost;
 pub use tx_filter::geyser_tx_involves_wallet;
